Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,11 +1,11 @@
 #include <stdio.h>
 int main(){
-    float a,b;
+    int a,b;
     char c;
-    scanf("%f %f %c",&a, &b, &c);
-    if(c=='+'){printf("%f",a+b);}
-    else if (c=='-'){printf("%f",a-b);}
-    else if(c=='*'){printf("%f",a*b);}
-    else if (c=='/'){printf("%f",a/b);}
+    scanf("%d %d %c",&a, &b, &c);
+    if(c=='+'){printf("%d",a+b);}
+    else if (c=='-'){printf("%d",a-b);}
+    else if(c=='*'){printf("%d",a*b);}
+    else if (c=='/'){printf("%d",a/b);}
     return 0;
 }
